[UPD] article footer wording update
Update the footer that shows up at the bottom of every article. Remove
Instagram, since it's private now. Clean up grammar/wording. Remove link
to the footer. I do have links in the footer, but it's not immediatley
clear that there's contact info in the footer. It should be easy enough
for people to find me online.