[ADD] travis & skeleton testing suite setup
This creates a skeleton for two test suites:
- unit
- endpoint

The endpoint tests are e2e/integration tests. I'm using Playwright to do
this. I added chrome and firefox devices for desktop and mobile
initially.

The unit tests are using uvu. The core svelte developers are using uvu
to test their code, so going with that since it will remain stable and
supported and I can look to the core code for examples.

The general travis process at this point looks like:
- Ubuntu bionic container
- Node 15
- Notifies on failure
- Includes Playwright apt dependencies
- Runs tests using `npm test`