[FIX] article meta loading issue
Articles were originally dynamically loading in meta tags after the
component mounted. This prevented some crawlers from ever seeing the
meta details. I stripped that out of the onMount so that the SEO
component renders before the page finishes loading.

Added an e2e test to check that all articles are loading with 200's to
catch potential regression issues on stuff like this.